Stable var caching. The values of non-dynamic ns-resolved (def'ed) vars appearing by name in code are fixed at fn entry point, and will be cached for subsequent calls until the vars change (e.g. via a new def). Access to unbound vars will not always throw an exception, and may instead return Unbound objects. These Unbound objects throw exceptions when invoked.
Note that modifications of var roots in a fn body will not be seen by the code compiled against those names in the same fn body. If you want to treat top-level vars as boxes, use #'

This change allows the overhead of placing things in vars and small wrapper fns to be substantially eliminated.